ml-service/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.4
pymongo>=4.5.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
requests>=2.31.0

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# NLP
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0

# Fraud Detection ML
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Additional utilities
scipy>=1.11.0
joblib>=1.3.0
